# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SunoSathi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/sessions:
    get:
      tags: [chat]
      operationId: listMyChatSessions
      summary: List my chat sessions
      responses:
        '200':
          description: Sessions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatSession'
    post:
      tags: [chat]
      operationId: startChatSession
      summary: Start a chat session with a listener (deducts ₹4/min from wallet on each minute)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartChatBody'
      responses:
        '200':
          description: Session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSession'
        '402':
          description: Insufficient wallet balance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /chat/sessions/{id}:
    get:
      tags: [chat]
      operationId: getChatSession
      summary: Get a chat session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSession'
  /chat/sessions/{id}/messages:
    get:
      tags: [chat]
      operationId: listChatMessages
      summary: List chat messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Messages.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessage'
    post:
      tags: [chat]
      operationId: sendChatMessage
      summary: Send a chat message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageBody'
      responses:
        '200':
          description: Message created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessage'
  /chat/sessions/{id}/end:
    post:
      tags: [chat]
      operationId: endChatSession
      summary: End a chat session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session ended.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSession'
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ChatSession:
      type: object
      required: [id, listenerId, listenerName, listenerPhotoUrl, userId, userName, status, kind, startedAt, endedAt, billedMinutes, totalCostInRupees, hasReview]
      properties:
        id:
          type: string
        listenerId:
          type: string
        listenerName:
          type: string
        listenerPhotoUrl:
          type: string
        userId:
          type: string
        userName:
          type: string
        status:
          type: string
          enum: [active, ended]
        kind:
          type: string
          enum: [chat, call]
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: [string, 'null']
          format: date-time
        billedMinutes:
          type: integer
        totalCostInRupees:
          type: integer
        lastMessagePreview:
          type: [string, 'null']
        hasReview:
          type: boolean
    StartChatBody:
      type: object
      required: [listenerId, kind]
      properties:
        listenerId:
          type: string
        kind:
          type: string
          enum: [chat, call]
    ChatMessage:
      type: object
      required: [id, sessionId, senderRole, body, createdAt]
      properties:
        id:
          type: string
        sessionId:
          type: string
        senderRole:
          type: string
          enum: [user, listener, system]
        body:
          type: string
        createdAt:
          type: string
          format: date-time
    SendMessageBody:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 1000
...